advanced-data/lambdas/requirements.txt
python-dotenv
pandas
awswrangler
boto3
python-dotenv